requirements.txt
Django==3.2.8
django-cors-headers==3.10.0
djangorestframework==3.12.4
pandas==1.3.4